requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-multipart
python-dotenv
email-validator
